refactoring controllers
Ho creato un controller separato e relativa route per il controller che recupera le categorie e il numero di prodotti per categoria
